services/consumer-analyticis/requirements.txt
pandas==2.3.3
requests==2.32.5
python-dotenv==1.2.1
waitress==3.0.2
Flask==3.1.2
plotly==6.4.0